pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smallevals"
version = "0.1.0"
description = "Small Language Models Evaluation Suite for RAG Systems"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mehmet Burak Sayıcı", email = "mburaksayici@gmail.com"}
]
keywords = ["rag", "evaluation", "vector-database", "retrieval", "nlp", "llm", "small-language-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core dependencies
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sentence-transformers>=5.1.2",
    "pandas>=2.0.0",
    "llama-cpp-python>=0.3.16",
    "peft>=0.18.0",
    "hf-transfer>=0.1.9",
    "langchain-community>=0.4.1",
    "llama-index-core>=0.10.0",
    # UI/Dashboard
    "plotly>=5.0.0",
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    # All Vector Databases
    "chromadb>=1.3.5",
    "weaviate-client>=4.18.1",
    "faiss-cpu>=1.13.0",
    "pinecone-client>=3.0.0",
    "pymilvus>=2.6.3",
    "elasticsearch==8.11",
    "elastic-transport==8.11",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.0",
    "qdrant-client>=1.16.1",
    "pymongo>=4.0.0",
    "turbopuffer>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/mburaksayici/smallevals"
Repository = "https://github.com/mburaksayici/smallevals"
Issues = "https://github.com/mburaksayici/smallevals/issues"
Documentation = "https://github.com/mburaksayici/smallevals#readme"

[project.optional-dependencies]
# Testing only
test = [
    "pytest>=9.0.1",
    "pytest-cov>=4.0.0",
    "testcontainers[elasticsearch,weaviate]>=4.13.3",
    "pyarrow>=10.0.0",
]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["smallevals"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[project.scripts]
smallevals = "smallevals.cli:main"
dash = "smallevals.ui_dash.app:main"
